Revert "Added nonce & implemented GTM and CSP changes  (#1661)"
This reverts commit 81b7d8c3aa2a77728c0aa7b59c052ac6cad51146.
