+++ js/analytics.js
@@ -50,7 +50,25 @@ if (checkDoNotTrack() === false) {
         return identifier;
       };
 
-      ReactGA.initialize(getIdentifier(`ga`));
+      const GA_ID = getIdentifier(`ga`);
+      const GTM_ID = getIdentifier(`gtm`);
+
+      if (GA_ID) {
+        ReactGA.initialize(GA_ID);
+      }
+
+      if (GTM_ID) {
+        (function (w, d, s, l, i) {
+          w[l] = w[l] || [];
+          w[l].push({ "gtm.start": new Date().getTime(), event: "gtm.js" });
+          var f = d.getElementsByTagName(s)[0],
+            j = d.createElement(s),
+            dl = l != "dataLayer" ? "&l=" + l : "";
+          j.async = true;
+          j.src = "https://www.googletagmanager.com/gtm.js?id=" + i + dl;
+          f.parentNode.insertBefore(j, f);
+        })(window, document, "script", "dataLayer", GTM_ID);
+      }
     },
   };
 }
